Resolve #557, init delete chart support
